sources/requirements.txt
appdirs==1.4.3
beautifulsoup4==4.5.3
coreapi==2.3.0
coreschema==0.0.4
Django==1.11
django-filter==1.0.2
django-nose==1.4.4
django-redis==4.8.0
django-rest-swagger==2.1.2
djangorestframework==3.6.2
freeze==1.0.10
itypes==1.1.0
Jinja2==2.9.6
Markdown==2.6.8
MarkupSafe==1.0
natsort==5.0.3
nose==1.3.7
openapi-codec==1.3.1
packaging==16.8
psycopg2==2.7.1
pyparsing==2.2.0
pytz==2017.2
redis==2.10.5
requests==2.13.0
simplejson==3.10.0
six==1.10.0
uritemplate==3.0.0
